Encode path into latin_1 bytes string
python-hglib expects all strings as bytes strings for input and
output. In Python 2, this was automatic, but Python 3+ needs this
expressed explicitly. Otherwise, for example, repo.log() throws an
error because the files parameter is passed as an str string but
among other things checks for null characters in bytes strings,
thus dealing with a type mismatch. encode() the string with the
same codec as python-hglib's b() from util.py.

As Python 2 is no longer supported, make the encode() unconditional.